backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy>=2
aiosqlite
python-dotenv